size_t -> int32_t

+++ src/NetBox/Far3Storage.cpp
@@ -224,8 +224,8 @@ UnicodeString TFar3Storage::DoReadStringRaw(const UnicodeString & Name, const Un
   return Result;
 }
 
-size_t TFar3Storage::DoReadBinaryData(const UnicodeString & Name,
-  void * Buffer, size_t Size)
+int32_t TFar3Storage::DoReadBinaryData(const UnicodeString & Name,
+  void * Buffer, int32_t Size)
 {
   return FPluginSettings.Get(FRoot, Name.c_str(), Buffer, Size);
 }
@@ -251,7 +251,7 @@ void TFar3Storage::DoWriteInt64(const UnicodeString & Name, int64_t Value)
 }
 
 void TFar3Storage::DoWriteBinaryData(const UnicodeString & Name,
-  const void * Buffer, size_t Size)
+  const void * Buffer, int32_t Size)
 {
   FPluginSettings.Set(FRoot, Name.c_str(), Buffer, Size);
 }
